--- shippable.yml
@@ -1,10 +1,0 @@
-language: go
-go:
-  - 1.11.1
-build:
-  ci:
-    - which go
-    - go version
-    - ls
-    - diff -u <(echo -n) <(gofmt -d ./)
-    - go get -u github.com/julienschmidt/httprouter
