+++ steamapi.go
@@ -1,12 +1,18 @@
-package mockapi
+package steamapi
 
 import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"net/http"
+	"net/url"
+	"strings"
 )
 
 const dataLocation = "internal/pkg/mock/mockdata"
+const baseURL = "https://api.steampowered.com"
+const userService = "ISteamUser"
+const playerService = "IPlayerService"
 
 // Player contains details about the Steam User
 type Player struct {
@@ -62,26 +68,63 @@ type FriendsResult struct {
 	Response FriendsList `json:"response"`
 }
 
-// Service is the type that owns methods for fetching mock steam data
+// Service is the type that owns methods for fetching steam data
 type Service struct {
+	client *http.Client
+	apiKey string
+}
+
+// NewService returns a service object configured with the provided Steam web API Key
+func NewService(apiKey string) *Service {
+	return &Service{
+		client: &http.Client{},
+		apiKey: apiKey,
+	}
 }
 
 // Players accepts one or more steamIDs and returns a PlayersResult
-func (m *Service) Players(steamID string) (*PlayersResult, error) {
-	rawPlayers, err := ioutil.ReadFile(dataLocation + "/players.json")
+func (s *Service) Players(steamIDs []string) (*PlayersResult, error) {
+	fmt.Printf("starting steam PLAYERS call for steamID list: %s\n", steamIDs)
+	req, err := http.NewRequest(
+		http.MethodGet,
+		baseURL+"/"+userService+"/GetPlayerSummaries/v0002?"+url.Values{
+			"steamids": {strings.Join(steamIDs, ",")},
+			"key":      {s.apiKey},
+		}.Encode(),
+		nil,
+	)
 	if err != nil {
+		fmt.Println("got an error forming the request")
 		return nil, err
 	}
-	var parsedPlayers PlayersResult
-	parseErr := json.Unmarshal(rawPlayers, &parsedPlayers)
-	if parseErr != nil {
-		return nil, parseErr
+	fmt.Println("successfully formed the http request")
+	req.Header.Add("Content-Type", "application/json")
+	fmt.Println("just added the content type header")
+	res, err := s.client.Do(req)
+	if res != nil {
+		fmt.Println("got a response, so defering the body close")
+		defer res.Body.Close()
 	}
+
+	if err != nil {
+		fmt.Println("got an error on .Do of http request")
+		return nil, err
+	}
+	fmt.Println("no error received in http request")
+	parsedPlayers := PlayersResult{}
+	// data, _ := ioutil.ReadAll(res.Body)
+	// fmt.Println(string(data))
+	if err := json.NewDecoder(res.Body).Decode(&parsedPlayers); err != nil {
+		fmt.Println("got an error decoding the http response body")
+		fmt.Printf("the err is: %s\n", err)
+		return nil, err
+	}
+	fmt.Println("no errors received in PLAYERS steam call")
 	return &parsedPlayers, nil
 }
 
 // Player accepts one steamID and returns that player's object
-func (m *Service) Player(steamID string) (*Player, error) {
+func (s *Service) Player(steamID string) (*Player, error) {
 	rawPlayers, err := ioutil.ReadFile(dataLocation + "/players.json")
 	if err != nil {
 		return nil, err
@@ -101,7 +144,7 @@ func (m *Service) Player(steamID string) (*Player, error) {
 }
 
 // Games accepts one or more steamIDs and returns a GamesResult
-func (m *Service) Games(steamIDs string) (*GamesResult, error) {
+func (s *Service) Games(steamIDs string) (*GamesResult, error) {
 	rawGames, err := ioutil.ReadFile(dataLocation + "/games.json")
 	if err != nil {
 		return nil, err
@@ -115,7 +158,7 @@ func (m *Service) Games(steamIDs string) (*GamesResult, error) {
 }
 
 // Friends accepts a steamID and returns all friends for that ID
-func (m *Service) Friends(steamIDs string) (*FriendsResult, error) {
+func (s *Service) Friends(steamIDs string) (*FriendsResult, error) {
 	rawFriends, err := ioutil.ReadFile(dataLocation + "/friends.json")
 	if err != nil {
 		return nil, err
